Squashed commit of the following:

commit e3d95f7204588bfe865d68c115f5eafe2e0fc1b3
Merge: f2c3a57 16362de
Author: Lyy <liuyuanyuan02@agora.io>
Date:   Wed Jun 22 16:20:03 2022 +0800

    Merge branch 'release/2.6.0' into hotfix/2.6.0_fix

commit f2c3a5754e0074a90345bc6860cf71e4b3b79216
Author: Lyy <liuyuanyuan02@agora.io>
Date:   Wed Jun 22 15:51:04 2022 +0800

    [FIX]:APAAS-3012 save image

commit 60357772cf34a5efc19700fde3ff9b6d55a9b579
Author: Lyy <liuyuanyuan02@agora.io>
Date:   Wed Jun 22 15:22:23 2022 +0800

    [FIX]:APAAS-3035 send initial granted users

commit 4c3727f0980e92f3fd45026c72f99d26240a026e
Author: Lyy <liuyuanyuan02@agora.io>
Date:   Wed Jun 22 12:21:31 2022 +0800

    [FIX]:APAAS-3032 add shape name

+++ AgoraWidgets/Whiteboard/FcrBoardWidget.swift
@@ -382,6 +382,7 @@ private extension FcrBoardWidget {
             guard let `self` = self else {
                 return
             }
+            self.dt.imageCountToSave = imagePathList.count
             for path in imagePathList {
                 guard let image = UIImage(contentsOfFile: path) else {
                     continue
@@ -397,18 +398,22 @@ private extension FcrBoardWidget {
     @objc func didFinishSavingImage(image: UIImage,
                                     error: NSError?,
                                     contextInfo: UnsafeRawPointer?) {
-        try? FileManager.default.removeItem(atPath: dt.currentSnapshotFolder)
-        AgoraWidgetLoading.removeLoading(in: view)
-        
         if let error = error {
-          log(content: "[FcrBoardWidget]: didFinishSavingImage error",
+          log(content: "[FcrBoardWidget]: save single image error",
               extra: error.description,
               type: .error)
+        } else {
+            log(content: "[FcrBoardWidget]: save single image successfully",
+                extra: dt.currentSnapshotFolder,
+                type: .info)
+        }
+        self.dt.imageCountToSave -= 1
+        guard self.dt.imageCountToSave == 0 else {
+            return
         }
 
-        log(content: "[FcrBoardWidget]: didFinishSavingImage",
-            extra: dt.currentSnapshotFolder,
-            type: .info)
+        try? FileManager.default.removeItem(atPath: dt.currentSnapshotFolder)
+        AgoraWidgetLoading.removeLoading(in: view)
         
         self.sendMessage(signal: .OnBoardSaveResult(.savedToAlbum))
       }
@@ -437,8 +442,9 @@ private extension FcrBoardWidget {
         if let usageExtra = info.roomProperties?.toObj(FcrBooardUsageOfExtra.self) {
             dt.grantedUsers = usageExtra.grantedUsers
         }
-        
+        // 发送页数，白板授权用户初始状态
         sendMessage(signal: .BoardPageChanged(.count(dt.page.count)))
+        sendMessage(signal: .GetBoardGrantedUsers(Array(dt.grantedUsers.keys)))
     }
 }
